requirements.txt
# Requirements per JBK Gestione con Supabase
# Dipendenze core
toga>=0.4.0
pillow

# Dipendenze Supabase
supabase>=2.3.4
python-dotenv>=1.0.0

# Dipendenze per database e autenticazione
psycopg2-binary>=2.9.9
python-dateutil>=2.8.2

# Dipendenze per gestione password e sicurezza
bcrypt>=4.1.2
pyjwt>=2.8.0

# Dipendenze per API e networking
httpx>=0.27.0
asyncio

# Dipendenze per testing (opzionali)
pytest>=7.4.4
pytest-asyncio>=0.23.5

# Dipendenze per development (opzionali)
black>=24.1.1
flake8>=7.0.0